some bugfixed

+++ server/src/main/java/com/restapi/server/model/FoodContainer.java
@@ -28,9 +28,9 @@ public class FoodContainer {
     @Column
     private String type;
     @Column
-    private long lng;
+    private double lng;
     @Column
-    private long lat;
+    private double lat;
     @Column
     private String address;
     @Column
